[FEAT][LG] - RestController & service de Consumer + Order attente de validation

+++ Back End/Projet E-Commerce/src/main/java/com/example/ecommerce/repository/IConsumerRepository.java
@@ -3,7 +3,7 @@ package com.example.ecommerce.repository;
 import com.example.ecommerce.entity.Consumer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IConsumerRepository extends JpaRepository<Consumer, Integer> {
+public interface IConsumerRepository extends JpaRepository<Consumer, Long> {
 
     Consumer findConsumerByEmailAndPassword(String email, String password);
